third_party/blink/web_tests/fast/events/page-visibility-prefixed-expected.txt

This test checks that Page Visibility state values are correct and the event changes are fired correctly.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS document.webkitVisibilityState is "visible"
PASS document.webkitHidden is false
PASS document.webkitVisibilityState is "hidden"
PASS document.webkitHidden is true
PASS document.webkitVisibilityState is "hidden"
PASS document.webkitHidden is true
PASS document.webkitVisibilityState is "visible"
PASS document.webkitHidden is false
PASS successfullyParsed is true

TEST COMPLETE
